Cleanup: move direct runner test to correct location.
